cast off_t as (unsigned long) so that tests build on MacOS 10.8. MacOS and FreeBSD should both serve as a useful reference for the correctness of the tests.
git-svn-id: svn://svn.code.sf.net/p/libkqueue/code/trunk@651 fb4e3144-bc1c-4b72-a658-5bcd248dd7f7

+++ test/read.c
@@ -398,7 +398,7 @@ test_kevent_regular_file(struct test_context *ctx)
     ret.data--;
     if ((curpos = lseek(fd, ret.data, SEEK_SET)) != ret.data) {
         printf("seek to %u failed with rv=%lu\n", 
-                (unsigned int) ret.data, curpos);
+                (unsigned int) ret.data, (unsigned long) curpos);
         abort();
     }
 
@@ -407,7 +407,7 @@ test_kevent_regular_file(struct test_context *ctx)
     ret.data = curpos + 1;
     if ((curpos = lseek(fd, ret.data, SEEK_SET)) != ret.data) {
         printf("seek to %u failed with rv=%lu\n",
-                (unsigned int) ret.data, curpos);
+                (unsigned int) ret.data, (unsigned long) curpos);
         abort();
     }
